+++ app/src/main/java/com/example/siki/activities/PromotionDetailActivity.java
@@ -44,7 +44,7 @@ public class PromotionDetailActivity extends AppCompatActivity {
     Button btnBack, btnSua, btnAnhKM, btnNgayBDKM, btnNgayKTKM;
     ImageView imgAnhKM, ivAnhLoaiSp;
     Spinner spLoaiSp;
-    Long idCategory;
+    Integer idCategory;
     String imagePath;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -150,7 +150,7 @@ public class PromotionDetailActivity extends AppCompatActivity {
 
         }
 
-        Map<Long, String> listCategory = categoryDatabase.getAllCategory();
+        Map<Integer, String> listCategory = categoryDatabase.getAllCategory();
         List<String> listLoaiSp = new ArrayList<>(listCategory.values()) ;
         listLoaiSp.add(0, "Chọn loại sản phẩm");
         ArrayAdapter<String> loaiSpAdapter = new ArrayAdapter<>(this,
@@ -162,7 +162,7 @@ public class PromotionDetailActivity extends AppCompatActivity {
             @Override
             public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
                 String selectedItem = (String) parent.getItemAtPosition(position);
-                for (Map.Entry<Long, String> entry : listCategory.entrySet()) {
+                for (Map.Entry<Integer, String> entry : listCategory.entrySet()) {
                     if (entry.getValue().equals(selectedItem)) {
                         idCategory = entry.getKey();
                         Picasso.get().load(categoryDatabase.findImagePathById(Math.toIntExact(entry.getKey()))).into(ivAnhLoaiSp);
